initial refactor of frontend
