+++ cosc360-rsvp/client/src/app/App.jsx
@@ -1,13 +1,13 @@
 import { Routes, Route } from "react-router-dom";
 import './css/App.css'
-import RegisterForm from './components/RegisterForm.jsx'
-import Homepage from "./pages/home.jsx";
-import AdminUserManage from './pages/AdminUserManage.jsx'
-import EventPage from "./pages/event.jsx";
+import RegisterForm from '../client/components/RegisterForm.jsx'
+import Homepage from "../../../src/pages/home.jsx";
+import AdminUserManage from '../features/adminManagement/AdminUserManage.jsx'
+import EventPage from "../features/event/event.jsx";
 
-import MyEvents from "./pages/myEvents.jsx";
-import SavedEvents from "./pages/savedEvents.jsx";
-import Login from "./pages/Login.jsx";
+import MyEvents from "../../../src/pages/myEvents.jsx";
+import SavedEvents from "../../../src/pages/savedEvents.jsx";
+import Login from "../features/login/Login.jsx";
 
 
 /* Create a page for each main view and then link it here! It will be accessible at localhost/pageName */
