+++ cosc360-rsvp/client/src/components/assets/cog.png
