+++ cosc360-rsvp/client/src/components/assets/trash-2.png
