+++ cosc360-rsvp/client/src/features/event/EventDetails.jsx
