+++ cosc360-rsvp/client/src/features/event/event_cards/EventCard.jsx
