+++ cosc360-rsvp/server/server.js
@@ -1,5 +1,5 @@
 import express from "express";
-import loginRouter from "../src/components/login/processLogin.js";
+import loginRouter from "../client/src/features/login/processLogin.js";
 import cors from "cors";
 import fs from "fs/promises";
 import search_events from "../src/services/search.js"
